requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc-getstream==1.13.0.post1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==14.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.5
dataclasses-json==0.6.7
decart==0.0.8
deepgram-sdk==6.0.1
distro==1.9.0
dnspython==2.8.0
docutils==0.22.4
elevenlabs==2.37.0
fastapi==0.134.0
flatbuffers==25.12.19
frozenlist==1.8.0
getstream==2.7.1
google-auth==2.48.0
google-crc32c==1.8.0
google-genai==1.65.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
id==1.6.1
idna==3.11
ifaddr==0.2.0
ijson==3.5.0
importlib-metadata==8.7.1
jaraco-classes==3.4.0
jaraco-context==6.1.0
jaraco-functools==4.4.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
markdown-it-py==4.0.0
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy-extensions==1.1.0
nh3==0.3.3
numpy==2.2.6
onnxruntime==1.24.2
opencv-python==4.13.0.92
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pillow==12.1.1
ping3==5.1.5
propcache==0.4.1
protobuf==7.34.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyee==13.0.1
pygments==2.19.2
pyjwt==2.11.0
pylibsrtp==1.0.0
pyopenssl==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32==311
pywin32-ctypes==0.2.3
readme-renderer==44.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
sse-starlette==3.3.2
starlette==0.52.1
structlog==23.3.0
supadata==1.6.0
sympy==1.14.0
tenacity==9.1.4
twine==6.2.0
twirp==0.0.7
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.41.0
vision-agents==0.3.8
vision-agents-plugins-decart==0.3.8
vision-agents-plugins-deepgram==0.3.8
vision-agents-plugins-elevenlabs==0.3.8
vision-agents-plugins-gemini==0.3.8
vision-agents-plugins-getstream==0.3.8
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
yt-dlp==2026.2.21
zipp==3.23.0
